Alignment issues seem to be related to mingw
It sounds like mingw is unable to do a 32-byte aligned stack on windows?
Or at least it can't with debugging symbols so maybe that explains why
the debug build crashes but the release build doesn't? MSVC seems to be
able to handle it on debug and release builds so maybe I'll just debug
on windows in VS