Working on AA issues
We should be shading misses as well but just coloring them by the
background color, otherwise edges of objects where we sometimes hit and
sometimes miss and hit nothing don't get anti-aliased. However this
still doesn't look to be quite enough as the quality isn't as nice as
what I get out of tray with 8spp.