Add support for taking more than 8 samples per pixel.
I'm not sure if I want to bother supporting < 8 samples per pixel,
especially since 8spp is really fast. Even with path tracing 8spp is
really not very long. So it might be an ok baseline. Also need to
compare the render quality against tray to see if things are being
sampled properly.